channel-liquidity-manager/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
influxdb-client>=1.36.0
flask>=3.0.0
flask-cors>=4.0.0
